Ctls and validation

See merge request swiftyteam/swifty!64

+++ src/gate/fdmdat.go
@@ -13,7 +13,7 @@ var fdmd sync.Map
 var tdat sync.Map
 
 type FnMemData struct {
-	mem	uint64
+	mem	uint
 	depname	string
 	fnid	string
 	ac	*AuthCtx
